add GrammarInjectorTest
